auth on keeps and vaults i think

+++ Controllers/KeepsController.cs
@@ -1,11 +1,14 @@
 using System;
 using System.Collections.Generic;
+using System.Security.Claims;
 using keepr.Models;
 using keepr.Repositories;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace keepr.Controllers
 {
+
   [Route("api/[controller]")]
   [ApiController]
   public class KeepsController : ControllerBase
@@ -43,13 +46,14 @@ namespace keepr.Controllers
         return BadRequest(e.Message);
       }
     }
-
+    [Authorize]
     // post, might need to refactor for posting with userId?
     [HttpPost]
     public ActionResult<Keep> Post([FromBody]Keep keep)
     {
       try
       {
+        keep.UserId = HttpContext.User.FindFirstValue("Id");
         return Ok(_repo.CreateKeep(keep));
       }
       catch (Exception e)
@@ -57,19 +61,26 @@ namespace keepr.Controllers
         return BadRequest("Bad Request");
       }
     }
+    [Authorize]
     // delete, might need to refactor to use userId
     [HttpDelete("{id}")]
     public ActionResult<string> Delete(int id)
     {
-      try
+      string userId = HttpContext.User.FindFirstValue("Id");
+      if (userId.Contains(userId))
       {
-        _repo.DeleteKeep(id);
-        return Ok("Successfully Deleted");
+
+        try
+        {
+          _repo.DeleteKeep(id);
+        }
+        catch (Exception e)
+        {
+          return BadRequest("Delete failed");
+        }
       }
-      catch (Exception e)
-      {
-        return BadRequest("Delete failed");
-      }
+      return Ok("Successfully Deleted");
     }
+    // Need getKeepsByVaultId or something, but need to figure out table join first
   }
 }
